move tauri and rocket functions to other files
