+++ src-tauri/src/rocket_endpoints.rs
@@ -1,0 +1,13 @@
+use rocket::{fs::FileServer, Ignite, Rocket};
+
+pub async fn setup_server() -> Result<Rocket<Ignite>, rocket::Error> {
+    let rocket_config = rocket::Config::figment()
+        .merge(("address", "0.0.0.0"))
+        .merge(("port", 8000))
+        .merge(("workers", 4));
+
+    rocket::custom(rocket_config)
+        .mount("/", FileServer::from("../src"))
+        .ignite()
+        .await
+}
